+++ src/rust/bitbox02/src/lib.rs
@@ -16,7 +16,8 @@
 // This crate contains safe wrappers around C functions provided by bitbox02_sys.
 #![no_std]
 
-#[cfg(test)]
+#[cfg(any(test, feature = "c-unit-testing", feature = "simulator-graphical"))]
+#[allow(unused_imports)]
 #[macro_use]
 extern crate std;
 
@@ -30,9 +31,13 @@ extern crate alloc;
 
 use alloc::string::String;
 
+#[cfg(feature = "bindings")]
+pub use bitbox02_sys as bindings;
+
 #[cfg(feature = "testing")]
 pub mod testing;
 
+pub mod delay;
 pub mod keystore;
 pub mod memory;
 pub mod random;
@@ -62,8 +67,8 @@ pub fn ug_put_string(x: i16, y: i16, input: &str, inverted: bool) {
     }
 }
 
-pub fn ug_clear_buffer() {
-    unsafe { bitbox02_sys::UG_ClearBuffer() }
+pub fn screen_clear() {
+    unsafe { bitbox02_sys::screen_clear() }
 }
 
 pub fn ug_send_buffer() {
@@ -76,6 +81,10 @@ pub fn ug_font_select_9x9() {
 
 pub fn ug_font_select_11x10() {
     unsafe { bitbox02_sys::UG_FontSelect(&bitbox02_sys::font_font_a_11X10) }
+}
+
+pub fn screen_rotate() {
+    unsafe { bitbox02_sys::screen_rotate() }
 }
 
 #[cfg_attr(not(target_arch = "arm"), allow(unused_variables))]