+++ CHANGELOG.md
@@ -7,6 +7,7 @@ customers cannot upgrade their bootloader, its changes are recorded separately.
 ## Firmware
 
 ### [Unreleased]
+- simulator-graphical: a new simulator with a graphical user interface
 
 ### v9.24.0
 - Change title when entering recovery words to `1 of 24`, `2 of 24`, etc.